+++ CHANGELOG.md
@@ -67,6 +67,8 @@
 42. Added MelonPreferences_Category.SetFilePath.
 43. Added MelonPreferences_Category.ResetFilePath.
 44. Fixed Issue with Core Initialization.
+45. Fixed Conflict Issue with Multiple MelonPreferences Categories using the same file.
+46. Fixed Issue with File Watcher System and Separated MelonPreferences Files.
 
 ---